// apps/web/src/types/ApiError.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ValidationField } from "./ValidationField";

export type ApiError = { "type": "NotFound", "data": { message: string, } } | { "type": "Unauthorized", "data": { message: string, } } | { "type": "Forbidden", "data": { message: string, } } | { "type": "ValidationError", "data": { fields: Array<ValidationField>, } } | { "type": "DatabaseError", "data": { message: string, } } | { "type": "InternalError", "data": { message: string, } } | { "type": "BadRequest", "data": { message: string, } } | { "type": "Conflict", "data": { message: string, } };